Implement Universal Summarizer
This is related to https://kagifeedback.org/d/1341-summarize-and-chat-with-any-webpage-via-a-browser-extension

I've got the whole thing to work for Chrome with a very basic UX. Before I spend the rest of the time porting to Firefox, I'd like to make sure this is acceptable.

The DX (Developer Experience) to build this is quite slow (every change requires rebuilding, extracting, and reloading), and there's no real standard in code style, linting, or tests, to make sure I got all scenarios covered, so I couldn't spend much time making the UX much better.

Ideally we'd have kind of a tab mechanism/UI to separate between preferences and functionality, or move the current popup to the extension options/preferences, for example, and just leave the new summarizer functionality as the new popup.

In any case, the time constraints can only get us basically this, with minor tweaks.