Firefox: Reduce scope of permissions request.
The universal summarizer only needs access to the active tab, and only
when started from a keyboard shortcut or a control in the extension UI.
Remove the `tabs` permission which allows background access to all tabs,
and replace it with the `activeTab` permission. Firefox does not display
a permissions prompt for `activeTab`.

The `declarativeNetRequest` functionality is only used to add an
Authorization header to requests made to kagi.com. In this case, access
to all requests is unnecessary, and the
`declarativeNetRequestWithHostAccess` permission can be used instead,
which gives access only to the domains listed in the manifest
`host_permissions`.