Align the manifest host_permission and declarativeNetRequest
The urlFilter used in the declarative net request rule was more broad
than the host_permissions field, which could potentially cause problems
with a permissions mismatch. Ensure that they're the same.

Switching the rule that adds the Authorization header to run on https
only is a good idea on its own; it ensures that the Authorization header
will not accidentally leak on an http request.