Reduce scope of permissions request in Firefox manifest (#15)
* Firefox: Reduce scope of permissions request.

The universal summarizer only needs access to the active tab, and only
when started from a keyboard shortcut or a control in the extension UI.
Remove the `tabs` permission which allows background access to all tabs,
and replace it with the `activeTab` permission. Firefox does not display
a permissions prompt for `activeTab`.

The `declarativeNetRequest` functionality is only used to add an
Authorization header to requests made to kagi.com. In this case, access
to all requests is unnecessary, and the
`declarativeNetRequestWithHostAccess` permission can be used instead,
which gives access only to the domains listed in the manifest
`host_permissions`.

* Align the manifest host_permission and declarativeNetRequest

The urlFilter used in the declarative net request rule was more broad
than the host_permissions field, which could potentially cause problems
with a permissions mismatch. Ensure that they're the same.

Switching the rule that adds the Authorization header to run on https
only is a good idea on its own; it ensures that the Authorization header
will not accidentally leak on an http request.