Improve UX - Separate Summarize UI (#21)
* Improve UX - Separate Summarize UI

This implements the summarization in a separate window/UI/popup. It allows for easier reference, saving multiple summaries if you want, and also enables a shortcut to instantly summarize the current page!

When triggered from the shortcut, it will use the API key and engine if it's set, and the default params.

* Remove unnecessary log

* Change Chrome shortcut

Avoid conflicting with Bitwarden

* Remove default shortcuts.

Also, close extension popup when summarizing.