Firefox for Android Tweaks (#26)
Related to#3

Basically, I've confirmed that none of the pieces of functionality we use are currently available in Firefox for Android 118a01 (Nightly). I did, however, make it so that we show the extension popup more nicely, and a different message so there's no visual indication the extension doesn't work.

- The search engine option doesn't show up, as the manifest indicator isn't recognized.
- The incognito session doesn't work, as we can't inject the token into the request headers
- The summarizer doesn't work since we'd need access to the active tab.

When requesting the permission, the error that comes up is the same as mentioned in https://stackoverflow.com/questions/49051084/firefox-console-error-webextension-context-not-found which seems to still be unresolved in https://bugzilla.mozilla.org/show_bug.cgi?id=1447806

Given https://blog.mozilla.org/addons/2023/08/10/prepare-your-firefox-desktop-extension-for-the-upcoming-android-release/ I'd expect we can revisit this in a couple of months and hopefully have some more of the extension working, but for now, it just looks better.

I also made some tweaks for Firefox (desktop) and Chrome, as I was testing everything, for UI consistency.