Save Summary Type & Target Language
This implements saving of the summary type and target language choices so they can be used for the next summarization (including via shortcut).

It effectively closes the request at https://kagifeedback.org/d/2321-chrome-extension-summarizer-extension-always-uses-summary-instead-of-key-moments

It also updates the expected `node` and `npm` versions for development (current LTS) and the `nodemon` development dependency, to resolve a security vulnerability without having to use `package.json:overrides`.