Add FastGPT input and Small Web link - v0.5.0
This implements a FastGPT input and Small Web link (with support for adding more links in the future). Both only show up _after_ having a valid session.

Because of the FastGPT input, if you have a valid session, opening the popup will now automatically focus that input field, so you can start typing right away. Pressing enter opens the FastGPT window with the query prefilled.

Clicking "Small Web" simply opens a new tab with that URL.

This also implements a fail-safe for issues reported in https://kagifeedback.org/d/2234-kagi-extension-api-key-not-persistent-in-firefox-esr-on-kali-linux and https://kagifeedback.org/d/2090-kagi-keeps-saying-invalid-session-token-in-firefox-on-linux-after-setting-up-successfully-and-working-fine-for-a-while and https://kagifeedback.org/d/2748-firefox-default-search-engine-keeps-resetting and hopefully make those easier to self-fix.

Finally, some linter and formatter tweaks and fixes were also added here, and the dependencies updated.