Do not sync token if privacy consent is not yet accepted
Previously, the extension would immediately sync the token and save it
in storage even if the user had not accepted the privacy policy.

This would result in a weird state where the user would then accept the
privacy policy, which would set the token to `undefined` in storage.

Then things would not properly sync and you'd be left wondering what
went wrong.

Just from this one line change, it should fix all of that.
It also has the plus side of actually requiring the user to accept the
policy first before it works as was intended.