Change auth header to be x-kagi-authorization
Instead of using the standard Authorization header, we instead want to
use a custom one.
This is because the standard authorization header, when sent to some of
our subdomains (like help.kagi.com), actually break since its backed by
a GCS bucket which reads that header.

This should remove any conflicts, and we already support this header on
our side :)