Update `chrome_settings_overrides.search_provider.suggest_url` to kagisuggest.com
This is mainly to complement use with Privacy Pass, but is functionally a no-op. The main difference is that cookies will not be sent with requests to `kagisuggest.com`

I don't know if this will update the autosuggest configuration for existing users or not, but it's fine in either case.